+++ goalc/data_compiler/dir_tpages.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+void compile_dir_tpages(const std::string& filename);