+++ docs/markdown/progress-notes/changelog.md
@@ -171,3 +171,5 @@
 - Forward declared basics can be used in more places
 - You can now set a field which has a forward declared structure or basic type
 - `cdr` now returns an object of type `pair`.
+- `lambda`s can now be used inside of a static object definition.
+- Methods can now be `:replace`d to override their type from their parent. Use this with extreme care.